==> 02_data_preparation.ipynb <==
##### Imports

from scripts.data_preparation import *

Progress tracking

# track progress for pipeline processing
tqdm.pandas()

## Data Preparation

### Real data

# lightly processed real user data
df_pandora = pd.read_csv("../original_data/pandora/PANDORA_original.csv", encoding = "utf-8")

# original features
feats = df_pandora.columns.to_list()
print(f"Original features:\n{feats}\n")

# size, features and sample
print(f"Shape:\n{df_pandora.shape}\n")
df_pandora.head(1)

Empty comments check and removal

# checking for empty comments
print("Check for empty comments")
print(len(df_pandora[df_pandora["body"].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])]))
print("Check for NaN comments")
print(len(df_pandora[df_pandora["body"].isna()]))

# drop NaN comments
df_pandora = df_pandora[df_pandora["body"].notna()].reset_index(drop = True)

Data types check and casting

# unique dtypes for each column
for c in df_pandora.columns:
    print(f"Unique dtypes {df_pandora.loc[:, c].apply(type).value_counts()}\n")
    
# casting dtypes
df_pandora["author_flair_text"] = df_pandora["author_flair_text"].astype(str)
df_pandora["score"] = df_pandora["score"].astype(float)
df_pandora["word_count"] = df_pandora["word_count"].astype(int)

Comments features exploration

analyze_distinct(df_pandora)

Irrelevant features removal
- a posteriori, we don't consider neither: parent_id, score, id and subreddit, so we can remove them

# features to delete
to_del = ["author_flair_text", "downs", "subreddit_id", "link_id", "controversiality", "gilded", "ups", "word_count", "word_count_quoteless"]

# drop irrelevant columns
df = df_pandora.drop(columns = to_del)
df.head(1)

Time conversion

# convert time from utc to date and time of day features
convert_utc_time(df, "created_utc")

### Simulated data (ex-ante moderations)
- we analyze comments focusing on before moderation content, which we need

# lightly processed simulated (ex-ante) data
df_sim = pd.read_csv("../original_data/simulator/exante/SIMULATOR_exante_original.csv", encoding = "utf-8")

# original features
feats = df_sim.columns.to_list()
print(f"Original features:\n{feats}\n")

# size, features and sample
print(f"Shape:\n{df_sim.shape}\n")
df_sim.head(1)

Empty comments check and removal

# checking for empty comments
print("Check for empty comments")
print(len(df_sim[df_sim["b_content"].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])]))
print("Check for NaN comments")
print(len(df_sim[df_sim["b_content"].isna()]))

# drop NaN comments
df_sim = df_sim[df_sim["b_content"].notna()].reset_index(drop = True)

Data types check

# unique dtypes for each column
for c in df_sim.columns:
    print(f"Unique dtypes {df_sim.loc[:, c].apply(type).value_counts()}\n")

Comments features exploration

analyze_distinct(df_sim)

### Data formatting and standardization
- the goal is to process the simulated data to make it as similar as possible to the real users'.

# real data viz
df.head(1)

# simulated data viz
df_sim.head(1)

Features renaming

# rename features according to real data (as much as possible)
df_sim = df_sim.rename(columns = {"username": "author"})

df_sim.head(1)

Features rearrangement

df = df[["author", "body", "gender", "age", "openness", "conscientiousness", 
         "extraversion", "agreeableness", "neuroticism", "score", "subreddit",
         "id", "parent_id", "date", "time_of_day"]]

df_sim = df_sim[["author", "comment_id", "b_content", "a_content_ofsa", "a_content_neut", "a_content_emp", "a_content_pres", 
                 "gender", "age", "openness", "conscientiousness", "extraversion", "agreeableness", 
                 "neuroticism", "thread_id", "node_id", "parent_id", "root_id", "race", "income", "education", 
                 "sex_orientation", "political_leaning", "religion", "simulate_seed"]]

df.head(1)

df_sim.head(1)

Data representation

# homogeneous values across datasets (gender)
df_sim["gender"] = df_sim["gender"].apply(lambda x: "f" if x == "female" else "m")

##### Discretization of OCEAN features
- since the simulated data has OCEAN features discretized, we do the same in the real users' data

# bins
bins = [0, 20, 40, 60, 80, 100]
# labels
labels = ["very low", "low", "medium", "high", "very high"]

# discretization of OCEAN
ocean = ["openness", "conscientiousness", "extraversion", "agreeableness", "neuroticism"]

for c in ocean:
    df[c] = pd.cut(df[c], bins = bins, labels = labels, right = True, include_lowest = True)

df.head(1)

### Before moderation dataset
- we only consider the comments before moderation
- we remove information about all a_contents (for each moderation)
- we remove duplicates for b_content
- we remove missing values if any

# before moderation dataset
df_bef = df_sim.drop(columns = ["a_content_ofsa", "a_content_neut", "a_content_emp", "a_content_pres"])
# rename the b_content into body consistently to the real data
df_bef.rename(columns = {"b_content": "body"}, inplace = True)

print(f"Shape:\n{df_bef.shape}")
df_bef.head(1)

# remove duplicates and missing values of the content before moderation
process_dataset(df_bef, "body")

### After moderation datasets
- we only consider the comments after moderation, by type of moderation
- we remove information about the b_content, focusing on after moderation comments
- we remove duplicates for a_content, but based on the type of moderation
- we remove missing values if any

# after moderation datasets
df_ofsa = df_sim.drop(columns = ["b_content", "a_content_neut", "a_content_emp", "a_content_pres"])
df_neut = df_sim.drop(columns = ["b_content", "a_content_ofsa", "a_content_emp", "a_content_pres"])
df_emp = df_sim.drop(columns = ["b_content", "a_content_neut", "a_content_ofsa", "a_content_pres"])
df_pres = df_sim.drop(columns = ["b_content", "a_content_neut", "a_content_emp", "a_content_ofsa"])

# rename the a_content into body consistently to the real data
df_ofsa.rename(columns = {"a_content_ofsa": "body"}, inplace = True)
df_neut.rename(columns = {"a_content_neut": "body"}, inplace = True)
df_emp.rename(columns = {"a_content_emp": "body"}, inplace = True)
df_pres.rename(columns = {"a_content_pres": "body"}, inplace = True)

print(f"Shape ofsa:\n{df_ofsa.shape}")
print(f"Shape neut:\n{df_neut.shape}")
print(f"Shape emp:\n{df_emp.shape}")
print(f"Shape pres:\n{df_pres.shape}")
df_ofsa.head(1)

# processing datasets after moderations
print("ofsa")
process_dataset(df_ofsa, "body")

print("neutral")
process_dataset(df_neut, "body")

print("empathizing")
process_dataset(df_emp, "body")

print("prescriptive")
process_dataset(df_pres, "body")

## Comments pre-processing

General cleansing of textual data.
- html, urls, mentions, hashtags, replacement symbols, contractions extension, extra spaces removal

# real comments
general_text_cleaning(df, "body")

# simulated comments
# before moderation
general_text_cleaning(df_bef, "body")

# after moderation
after_datasets = [df_ofsa, df_neut, df_emp, df_pres]

for d in after_datasets:
    general_text_cleaning(d, "body")

Check empty real data

check_empty(df, "body")

Check empty simulated data

check_empty(df_bef, "body")
check_empty(df_ofsa, "body")
check_empty(df_neut, "body")
check_empty(df_emp, "body")
check_empty(df_pres, "body")

##### Separate features for different pipelines

- One comment column will be preprocessed lightly to make it suitable for LLMs.
- The other is processed to extract transparent (e.g. linguistic, affective, toxicity-related) features.

# duplicate comments column, one for llm pipeline
# real
df.insert(loc = 1, column = "llm_body", value = df["body"])
# simulated
df_bef.insert(loc = 2, column = "llm_body", value = df_bef["body"])
df_ofsa.insert(loc = 2, column = "llm_body", value = df_ofsa["body"])
df_neut.insert(loc = 2, column = "llm_body", value = df_neut["body"])
df_emp.insert(loc = 2, column = "llm_body", value = df_emp["body"])
df_pres.insert(loc = 2, column = "llm_body", value = df_pres["body"])

# rename comments column for feature extraction pipeline
# real
df = df.rename(columns = {"body":"std_body"})
# simulated
df_bef.rename(columns = {"body":"std_body"}, inplace = True)
df_ofsa.rename(columns = {"body":"std_body"}, inplace = True)
df_neut.rename(columns = {"body":"std_body"}, inplace = True)
df_emp.rename(columns = {"body":"std_body"}, inplace = True)
df_pres.rename(columns = {"body":"std_body"}, inplace = True)

# shapes
print(f"Shape real data:\n{df.shape}\n")

print(f"Shape simulated data before mod:\n{df_bef.shape}\n")
print(f"Shape simulated data after ofsa:\n{df_ofsa.shape}\n")
print(f"Shape simulated data after neutral:\n{df_neut.shape}\n")
print(f"Shape simulated data after empathizing:\n{df_emp.shape}\n")
print(f"Shape simulated data after prescriptive:\n{df_pres.shape}\n")

df.head(1)

df_bef.head(1)

### Storing the datasets

# store full real and simulated datasets
df.to_csv("../original_data/pandora/PANDORA.csv", index = False, encoding = "utf-8") 
df_bef.to_csv("../original_data/simulator/exante/SIMULATOR_exante_bef.csv", index = False, encoding = "utf-8")
df_ofsa.to_csv("../original_data/simulator/exante/SIMULATOR_exante_ofsa.csv", index = False, encoding = "utf-8")
df_neut.to_csv("../original_data/simulator/exante/SIMULATOR_exante_neut.csv", index = False, encoding = "utf-8")
df_emp.to_csv("../original_data/simulator/exante/SIMULATOR_exante_emp.csv", index = False, encoding = "utf-8")
df_pres.to_csv("../original_data/simulator/exante/SIMULATOR_exante_pres.csv", index = False, encoding = "utf-8")
