==> 03_feat_extr_pipe.ipynb <==
##### Imports

from scripts.feature_extraction import *
from scripts.plotting import *

Progress tracking

# track progress for pipeline processing
tqdm.pandas()

## Feature extraction pipeline (no toxicity)

### Real data

# load real users' data
df_real = pd.read_csv("../original_data/pandora/PANDORA.csv", encoding = "utf-8") 

# original features
print(f"Features:\n{df_real.columns.to_list()}")

# size, features and sample
print(f"Shape of dataset:\n{df_real.shape}\n")
df_real.head(1)

##### Pipeline on real comments

# feature with comments' content before moderation
text_col = "std_body"

# pipeline for processing text and extracting features
pipeline = [extract_counts, extract_emoji_counts, process_emojis, fix_encoding, 
            extract_emotions, extract_pol_subj, extract_VAD, extract_readability, 
            clean_text, lowercase, extract_word_counts, process_stopwords, lemmatization]

# execute pipeline on the comments before moderation
apply_pipeline(df_real, text_col, pipeline)

# visualization
print(f"Shape:\n{df_real.shape}")
df_real.head(1)

##### Post-processing

Missing values

# missing values for the processed text
print(f"Missing values for {text_col}:")
print(len(df_real[df_real[text_col].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])]))
print(len(df_real[df_real[text_col].isna()]))

# checking starting text to see why they are empty (should be, OK)
df_real[df_real[text_col].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])].iloc[:10, 1]

# removing empty comments and resetting index
df_clean = df_real[df_real[text_col].notna()].reset_index(drop = True)
df_clean = df_clean[~df_clean[text_col].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])].reset_index(drop = True)

print(f"Size before missing values removal: {df_real.shape}")
print(f"Size after missing values removal: {df_clean.shape}")

Missing emojis set (decided to keep them as empty strings)

# checking for empty emoji list and set
print("Missing emojis:")
print(len(df_clean[df_clean["emoji_unique"].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])]))
print(len(df_clean[df_clean["emoji_unique"].isna()]))

print("\nFound emojis:")
df_clean[~df_clean["emoji_unique"].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])]["emoji_unique"][:10]

Latin alphabet
- 339 comments were recognized as non-latin. Inspecting them, most were normal, so we decided to keep them

# apply function to the column
df_clean["is_latin"] = df_clean["llm_body"].progress_apply(is_latin)

# number of latin and non-latin comments
print(df_clean["is_latin"].value_counts())

# remove is_latin feature
df_clean.drop(columns = "is_latin", inplace = True)

# store the dataset
#df_clean.to_csv("../data/pandora/PANDORA_featextr.csv", index = False, encoding = "utf-8")

Reset kernel (RAM)

%reset -f

### Simulated data (ex-ante)

from scripts.feature_extraction import *
from scripts.plotting import *

# track progress for pipeline processing
tqdm.pandas()

# simulated data before moderation
df_bef = pd.read_csv("../original_data/simulator/exante/SIMULATOR_exante_bef.csv", encoding = "utf-8")
# ofsa
df_ofsa = pd.read_csv("../original_data/simulator/exante/SIMULATOR_exante_ofsa.csv", encoding = "utf-8")
# neutral
df_neut = pd.read_csv("../original_data/simulator/exante/SIMULATOR_exante_neut.csv", encoding = "utf-8")
# empathizing
df_emp = pd.read_csv("../original_data/simulator/exante/SIMULATOR_exante_emp.csv", encoding = "utf-8")
# prescriptive
df_pres = pd.read_csv("../original_data/simulator/exante/SIMULATOR_exante_pres.csv", encoding = "utf-8")

# drop NaN comments
df_bef.dropna(subset = ["std_body"], inplace = True)
df_ofsa.dropna(subset = ["std_body"], inplace = True)
df_neut.dropna(subset = ["std_body"], inplace = True)
df_emp.dropna(subset = ["std_body"], inplace = True)
df_pres.dropna(subset = ["std_body"], inplace = True)

# size, features and sample
print(f"Shape of dataset before mod:\n{df_bef.shape}\n")
print(f"Shape of dataset ofsa mod:\n{df_ofsa.shape}\n")
print(f"Shape of dataset neutral mod:\n{df_neut.shape}\n")
print(f"Shape of dataset empathizing mod:\n{df_emp.shape}\n")
print(f"Shape of dataset prescriptive mod:\n{df_pres.shape}\n")
df_bef.head(1)

##### Pipeline on content before moderation

# feature with comments' content before moderation
text_col = "std_body"

# pipeline for processing text and extracting features
pipeline = [extract_counts, extract_emoji_counts, process_emojis, fix_encoding, 
            extract_emotions, extract_pol_subj, extract_VAD, extract_readability, 
            clean_text, lowercase, extract_word_counts, process_stopwords, lemmatization]

# execute pipeline on the comments before moderation
apply_pipeline(df_bef, text_col, pipeline)

# visualization
print(f"Shape:\n{df_bef.shape}")
df_bef.head(1)

##### Post-processing

Missing values

# missing values for the processed text
print(f"Missing values for {text_col}:")
print(len(df_bef[df_bef[text_col].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])]))
print(len(df_bef[df_bef[text_col].isna()]))

# checking starting text to see why they are empty (should be, OK)
df_bef[df_bef[text_col].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])].iloc[:10, 1]

# removing empty comments and resetting index
df_clean = df_bef[df_bef[text_col].notna()].reset_index(drop = True)
df_clean = df_clean[~df_clean[text_col].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])].reset_index(drop = True)

print(f"Size before missing values removal: {df_bef.shape}")
print(f"Size after missing values removal: {df_clean.shape}")

Missing emojis set (decided to keep them as empty strings)

# checking for empty emoji list and set
print("Missing emojis:")
print(len(df_clean[df_clean["emoji_unique"].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])]))
print(len(df_clean[df_clean["emoji_unique"].isna()]))

print("\nFound emojis:")
df_clean[~df_clean["emoji_unique"].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])]["emoji_unique"][:10]

Latin alphabet

# apply function to the column
df_clean["is_latin"] = df_clean["llm_body"].progress_apply(is_latin)

# number of latin and non-latin comments
print(df_clean["is_latin"].value_counts())

# remove is_latin feature
df_clean.drop(columns = "is_latin", inplace = True)

# store the dataset
df_clean.to_csv("../data/simulator/exante/before_mod/SIMULATOR_exante_bef_featextr.csv", index = False, encoding = "utf-8")

##### Pipeline on content after OFSA

# execute pipeline on the comments before moderation
apply_pipeline(df_ofsa, text_col, pipeline)

# visualization
print(f"Shape:\n{df_ofsa.shape}")
df_ofsa.head(1)

##### Post-processing

Missing values

# missing values for the processed text
print(f"Missing values for {text_col}:")
print(len(df_ofsa[df_ofsa[text_col].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])]))
print(len(df_ofsa[df_ofsa[text_col].isna()]))

# checking starting text to see why they are empty (should be, OK)
df_ofsa[df_ofsa[text_col].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])].iloc[:10, 1]

# removing empty comments and resetting index
df_clean = df_ofsa[df_ofsa[text_col].notna()].reset_index(drop = True)
df_clean = df_clean[~df_clean[text_col].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])].reset_index(drop = True)

print(f"Size before missing values removal: {df_ofsa.shape}")
print(f"Size after missing values removal: {df_clean.shape}")

Missing emojis set (decided to keep them as empty strings)

# checking for empty emoji list and set
print("Missing emojis:")
print(len(df_clean[df_clean["emoji_unique"].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])]))
print(len(df_clean[df_clean["emoji_unique"].isna()]))

print("\nFound emojis:")
df_clean[~df_clean["emoji_unique"].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])]["emoji_unique"][:10]

Latin alphabet

# apply function to the column
df_clean["is_latin"] = df_clean["llm_body"].progress_apply(is_latin)

# number of latin and non-latin comments
print(df_clean["is_latin"].value_counts())

# remove is_latin feature
df_clean.drop(columns = "is_latin", inplace = True)

# store the dataset
df_clean.to_csv("../data/simulator/exante/after_mod/SIMULATOR_exante_ofsa_featextr.csv", index = False, encoding = "utf-8")

##### Pipeline on content after Neutral

# execute pipeline on the comments before moderation
apply_pipeline(df_neut, text_col, pipeline)

# visualization
print(f"Shape:\n{df_neut.shape}")
df_neut.head(1)

##### Post-processing

Missing values

# missing values for the processed text
print(f"Missing values for {text_col}:")
print(len(df_neut[df_neut[text_col].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])]))
print(len(df_neut[df_neut[text_col].isna()]))

# checking starting text to see why they are empty (should be, OK)
df_neut[df_neut[text_col].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])].iloc[:10, 1]

# removing empty comments and resetting index
df_clean = df_neut[df_neut[text_col].notna()].reset_index(drop = True)
df_clean = df_clean[~df_clean[text_col].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])].reset_index(drop = True)

print(f"Size before missing values removal: {df_neut.shape}")
print(f"Size after missing values removal: {df_clean.shape}")

Missing emojis set (decided to keep them as empty strings)

# checking for empty emoji list and set
print("Missing emojis:")
print(len(df_clean[df_clean["emoji_unique"].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])]))
print(len(df_clean[df_clean["emoji_unique"].isna()]))

print("\nFound emojis:")
df_clean[~df_clean["emoji_unique"].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])]["emoji_unique"][:10]

Latin alphabet

# apply function to the column
df_clean["is_latin"] = df_clean["llm_body"].progress_apply(is_latin)

# number of latin and non-latin comments
print(df_clean["is_latin"].value_counts())

# remove is_latin feature
df_clean.drop(columns = "is_latin", inplace = True)

# store the dataset
df_clean.to_csv("../data/simulator/exante/after_mod/SIMULATOR_exante_neut_featextr.csv", index = False, encoding = "utf-8")

##### Pipeline on content after Empathizing

# execute pipeline on the comments before moderation
apply_pipeline(df_emp, text_col, pipeline)

# visualization
print(f"Shape:\n{df_emp.shape}")
df_emp.head(1)

##### Post-processing

Missing values

# missing values for the processed text
print(f"Missing values for {text_col}:")
print(len(df_emp[df_emp[text_col].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])]))
print(len(df_emp[df_emp[text_col].isna()]))

# checking starting text to see why they are empty (should be, OK)
df_emp[df_emp[text_col].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])].iloc[:10, 1]

# removing empty comments and resetting index
df_clean = df_emp[df_emp[text_col].notna()].reset_index(drop = True)
df_clean = df_clean[~df_clean[text_col].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])].reset_index(drop = True)

print(f"Size before missing values removal: {df_emp.shape}")
print(f"Size after missing values removal: {df_clean.shape}")

Missing emojis set (decided to keep them as empty strings)

# checking for empty emoji list and set
print("Missing emojis:")
print(len(df_clean[df_clean["emoji_unique"].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])]))
print(len(df_clean[df_clean["emoji_unique"].isna()]))

print("\nFound emojis:")
df_clean[~df_clean["emoji_unique"].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])]["emoji_unique"][:10]

Latin alphabet

# apply function to the column
df_clean["is_latin"] = df_clean["llm_body"].progress_apply(is_latin)

# number of latin and non-latin comments
print(df_clean["is_latin"].value_counts())

# remove is_latin feature
df_clean.drop(columns = "is_latin", inplace = True)

# store the dataset
df_clean.to_csv("../data/simulator/exante/after_mod/SIMULATOR_exante_emp_featextr.csv", index = False, encoding = "utf-8")

##### Pipeline on content after Prescriptive

# execute pipeline on the comments before moderation
apply_pipeline(df_pres, text_col, pipeline)

# visualization
print(f"Shape:\n{df_pres.shape}")
df_pres.head(1)

##### Post-processing

Missing values

# missing values for the processed text
print(f"Missing values for {text_col}:")
print(len(df_pres[df_pres[text_col].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])]))
print(len(df_pres[df_pres[text_col].isna()]))

# checking starting text to see why they are empty (should be, OK)
df_pres[df_pres[text_col].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])].iloc[:10, 1]

# removing empty comments and resetting index
df_clean = df_pres[df_pres[text_col].notna()].reset_index(drop = True)
df_clean = df_clean[~df_clean[text_col].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])].reset_index(drop = True)

print(f"Size before missing values removal: {df_pres.shape}")
print(f"Size after missing values removal: {df_clean.shape}")

Missing emojis set (decided to keep them as empty strings)

# checking for empty emoji list and set
print("Missing emojis:")
print(len(df_clean[df_clean["emoji_unique"].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])]))
print(len(df_clean[df_clean["emoji_unique"].isna()]))

print("\nFound emojis:")
df_clean[~df_clean["emoji_unique"].isin(["", " ", "NaN", "None", "NULL", "null", "NA"])]["emoji_unique"][:10]

Latin alphabet

# apply function to the column
df_clean["is_latin"] = df_clean["llm_body"].progress_apply(is_latin)

# number of latin and non-latin comments
print(df_clean["is_latin"].value_counts())

# remove is_latin feature
df_clean.drop(columns = "is_latin", inplace = True)

# store the dataset
df_clean.to_csv("../data/simulator/exante/after_mod/SIMULATOR_exante_pres_featextr.csv", index = False, encoding = "utf-8")
