==> 08d_validation_prescriptive.ipynb <==
##### Imports

from scripts.validation import *
from scripts.plotting import * 

Progress tracking

# track progress
tqdm.pandas()

## Prescriptive moderation validation
- now that we have found toxicity-based clusters of comments, we want to validate the impact of moderations:
    - focusing on before moderation data against all moderation types
    - in terms of how well moderation strategies are able to reduce toxic behaviors
    - to do so, since the moderated comments are versions of the unmoderated comments
        - we look at how patterns change for each toxic profile
        - using the clusters found exlusively on unmoderated data

- we could do both global and local validations:
    - global, comparing full datasets against each other
    - local, comparing matching toxic profiles for each dataset

### Datasets

# simulated before mod
df_bef = pd.read_csv("../data_before_after/SIMULATOR_exante_bef_profiles.csv", encoding = "utf-8")
# moderated
df_pres = pd.read_csv("../data_before_after/SIMULATOR_exante_pres.csv", encoding = "utf-8")

##### Feature Engineering
- we want to derive features from the already extracted ones, could be more insightful

feat_engineering(df_pres)

##### Feature selection
- keep only features shared across all datasets (real and simulated), to be able to do comparative analyses
- we exclude also the textual bodies and other irrelevant features for the analysis, such as parent_id and age (which has many missing values in the real dataset)

# features to keep
feats_shared = [col for col in df_bef.columns if col in df_pres.columns]

df_pres = df_pres[feats_shared].copy()

print(f"Shape of before moderation data: {df_bef.shape}")
df_bef.head(2)

print(f"Shape of moderated data: {df_pres.shape}")
df_pres.head(2)

##### Features for each dimension to evaluate

## toxic dimension
tox_feat = ["obscene", "threat", "insult", "identity_attack"]

## personality dimension
ocean_feat = ["openness", "conscientiousness", "extraversion", "agreeableness", "neuroticism"]

## sentiment/emotional dimension
sent_feat = ["fear", "anger", "anticip", "trust", "surprise", "sadness", "disgust", "joy", "valence", "arousal", "dominance", "subjectivity", "sent_balance", "sent_emoji_balance"]
# not considered:
# ("positive" - "negative") -> "sent_balance"
# "polarity", reduntant with "sent_balance"
# ("num_emoji_pos" - "num_emoji_neg") -> "sent_emoji_balance"

## linguistic dimension
ling_feat = ["num_words", "num_words_unique", "punct_ratio", "upper_ratio", "emoji_ratio", "adj_ratio", "noun_ratio", "verb_ratio", "lex_ratio", "stopw_ratio", "words_sent_ratio", "ttr"]
# not considered:
# count_features -> count_feature / "num_words"
# "emoji_unique"
# "emoji_list"

## readability dimension
read_feat = ["flesch", "flesch_kincaid", "fog", "smog", "ari", "coleman_liau", "dale_chall", "linsear", "complex_ratio"]
# not considered:
# "difficult_words" -> "complex_ratio" = "difficult_words"/"num_words"

## Global Validation

equal_feats, diff_feats, valid_metrics = validation_metrics_mod(df_bef, df_pres, tox_cols = tox_feat, pers_cols = ocean_feat,
                                                            sent_cols = sent_feat, ling_cols = ling_feat, read_cols = read_feat,
                                                            pval_thr = 0.01, sort_by = "rel_mean_diff", top = None, name_1 = "Unmoderated", name_2 = "OSFA")

## Local Validation

### Profiles mapping
- to validate how patterns change between before and after moderation, we don't map the clusters
- we instead consider the profiles found for the unmoderated data
- and annotate the moderated comments with their unmoderated toxic profile
- we therefore evaluate how each toxic profile has been affected by the specific moderation type

# retrieve original toxic profiles for the moderated comments
df_pres = df_pres.merge(df_bef[["comment_id", "tox_profile"]], on = "comment_id", how = "left")

df_pres.head(2)

### Validation metrics

##### Healthy profile (no, we only consider toxic profiles)

#equal_neut, diff_neut, valid_metrics_neut = validation_metrics_mod(df_bef, df_pres, profile = "Healthy", tox_cols = tox_feat, pers_cols = ocean_feat, 
#                                                               sent_cols = sent_feat, ling_cols = ling_feat, read_cols = read_feat,
#                                                               pval_thr = 0.01, sort_by = "rel_mean_diff", top = None, name_1 = "Unmoderated", name_2 = "Prescriptive")

##### Vulgar profile

equal_vulg, diff_vulg, valid_metrics_vulg = validation_metrics_mod(df_bef, df_pres, profile = "Vulgar", tox_cols = tox_feat, pers_cols = ocean_feat, 
                                                               sent_cols = sent_feat, ling_cols = ling_feat, read_cols = read_feat,
                                                               pval_thr = 0.01, sort_by = "rel_mean_diff", top = None, name_1 = "Unmoderated", name_2 = "Prescriptive")

##### Offensive profile

equal_off, diff_off, valid_metrics_off = validation_metrics_mod(df_bef, df_pres, profile = "Offensive", tox_cols = tox_feat, pers_cols = ocean_feat, 
                                                               sent_cols = sent_feat, ling_cols = ling_feat, read_cols = read_feat,
                                                               pval_thr = 0.01, sort_by = "rel_mean_diff", top = None, name_1 = "Unmoderated", name_2 = "Prescriptive")

##### Discriminatory profile

equal_disc, diff_disc, valid_metrics_disc = validation_metrics_mod(df_bef, df_pres, profile = "Discriminatory", tox_cols = tox_feat, pers_cols = ocean_feat, 
                                                               sent_cols = sent_feat, ling_cols = ling_feat, read_cols = read_feat,
                                                               pval_thr = 0.01, sort_by = "rel_mean_diff", top = None, name_1 = "Unmoderated", name_2 = "Prescriptive")

##### Hostile profile

equal_host, diff_host, valid_metrics_host = validation_metrics_mod(df_bef, df_pres, profile = "Hostile", tox_cols = tox_feat, pers_cols = ocean_feat, 
                                                               sent_cols = sent_feat, ling_cols = ling_feat, read_cols = read_feat,
                                                               pval_thr = 0.01, sort_by = "rel_mean_diff", top = None, name_1 = "Unmoderated", name_2 = "Prescriptive")
